Fix: health_screen.dart 구조단순화
